Add Image.Extent() interface
Extent() return the next extent in the image, starting at the specified
offset, and limited by the specified length. An extent is a run of
clusters of with the same status.

Using Extent() we can iterate over the image and skip extents that reads
as zeros, instead of reading a zero buffer and detecting that the buffer
is full of zeros.

Benchmarking show 2 orders of magnitude speeded compared with zero
detection (10 TiB/s instead of 100 GiB/s)

    % go test -bench Benchmark0p/qcow2$/read
    Benchmark0p/qcow2/read-12      516     2503752 ns/op      107213.28 MB/s    1050537 B/op    39 allocs/op

    % go test -bench NextUnallocated
    BenchmarkNextUnallocated-12    116    10236186 ns/op    10489666.91 MB/s          0 B/op     0 allocs/op

Only qcow2 image has a real implementation. For other image formats we
treat all clusters as allocated, so Extent() always return one allocated
extent. For raw format we can implement Extents() later using SEEK_DATA
and SEEK_HOLE.

Extent() is more strict than ReadAt and fails when offset+length are
after the end of the image. We always know the length of the image so
requesting an extent after the end of the image is likely a bug in the
user code and failing fast may help to debug such issues.

Signed-off-by: Nir Soffer <nirsof@gmail.com>