Add qemuio test package
To qcow2 test images, we used raw image and converted to qcow2 format.
This does not work for creating qcow2 chain with some data in the
backing file and some data in the top image. qemu-io allows writing
data, zeros or discarding a byte range in an image of any format
supported by qemu.

This change add a tiny test package for running qemu-io. We provide 2
functions, wrapping usage of the write command with different flags:

- Write() - write a pattern to specified range
- Zero() - write zeros to specified range
- Discard() - discard specified range

With these functions we can test all kind of extents in both top layer
or the backing file:
- Allocated: false, Zero: true
- Allocated: true, Zero: false
- Allocated: true, Zero: true

Signed-off-by: Nir Soffer <nirsof@gmail.com>