Test image with a backing file
We cannot have useful tests for raw backing file since we have only a
stub implementation reporting that the entire backing file is allocated,
so we test only qcow2 image backing file.

We test backing file in qcow2 and qcow2 compressed. When both images are
uncompressed, allocated extents from top and base are merged. When base
is compressed, allocated extents from top and based cannot be merged,
since all base clusters are compressed.

Compressed image with backing file may be possible to crate with
qemu-io, but is not real use case so we don't test it.

Signed-off-by: Nir Soffer <nirsof@gmail.com>