Test short backing file
When adding a longer qcow2 image on top of an os image, we treat extents
after the end of the backing file as unallocated.

We test also the special case of raw backing file not aligned to cluster
size. When reading a short extent from the end of the backing file we
extend the length to full cluster.

Signed-off-by: Nir Soffer <nirsof@gmail.com>