Support convert progress
Previously we always read the entire image so we could report progress
by wrapping the ReaderAt so every read updates the progress indicator.
Now we read only the allocated extents, so we need another way.

To report progress, you need to pass now a type implementing the Updater
interface. We could also use a function but this way seems cleaner.

Signed-off-by: Nir Soffer <nirsof@gmail.com>