Add progress bar to example program
Most users will want to integrate a progress bar and this serve as a
good example of how to do this. This also makes it more fun to work on
this project.

Ubuntu 24.04 image in qcow2 compressed format:

    % ./go-qcow2reader-example convert /var/tmp/images/test.zlib.qcow2 /tmp/tmp.img
    6.00 GiB / 6.00 GiB [-------------------------------------] 100.00% 2.86 GiB p/s

Ubuntu 24.04 image in qcow2 format:

    % ./go-qcow2reader-example convert /var/tmp/images/test.qcow2 /tmp/tmp.img
    6.00 GiB / 6.00 GiB [------------------------------------] 100.00% 15.73 GiB p/s

16 TiB empty qcow2 image:

    % ./go-qcow2reader-example convert /var/tmp/images/test.0p.qcow2 /tmp/tmp.img
    16.00 TiB / 16.00 TiB [---------------------------------] 100.00% 152.46 TiB p/s

The pb module brings too many dependencies. Maybe we can find another
library or write a very simple progress instead.

Signed-off-by: Nir Soffer <nirsof@gmail.com>